requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu111
accelerate==1.1.1
aiofiles==23.2.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
autocommand==2.2.2
backports.tarfile==1.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
diffusers==0.31.0
einops==0.8.0
exceptiongroup==1.2.2
fastapi==0.115.6
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.55.2
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.25.2
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
inflect==7.3.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
Jinja2==3.1.4
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.3
mdurl==0.1.2
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.1
numpy==2.1.3
omegaconf==2.3.0
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.2.2
psutil==6.1.0
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruff==0.8.2
safehttpx==0.1.6
safetensors==0.4.5
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.13.2
torch==2.5.1+cu124
torchvision==0.20.1+cu124
tqdm==4.67.1
transformers==4.44.2
typeguard==4.3.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
websockets==14.1
zipp==3.21.0
